Fix #2502: Fix the compiler handling of Int shift Long.
Unlike other operations involving an Int and a Long, the specific
case of Int shift Long is an Int operation, and must behave as
such. In particular, the right hand side should be coerced to an
Int (instead of the left hand side being coerced to a Long).

This change is similar to what was done upstream in
https://github.com/scala/scala/commit/eb954cafa2737a6f06f8a802ceec133a88959f71