Fix #2513: Annotate `@JSExport` with `@field @getter @setter`.
In addition, the compiler and the `js.use(x).as[T]` macro need to
look for `@JSExport` on the getter/setter itself, if there is no
accessed field. In theory, it should do so unconditionally, but
backward binary compatibility demand that we still recognize
exports built with a previous version of the library.

Incidentally, this provides a better fix to the old issue #970.
It is not necessary to use `@(JSExport @meta.field)` for
constructor parameter fields anymore.