Fix #2497: Include or exclude list in HTML runner
Listing all the excluded tests in the URL might cause a request size
that is too large if most tests are excluded. Therefore, we create an
include or an exclude list, depending on the size of each respective
set.
Further, when creating an include list, we hash the excluded tests to
avoid implicitly excluding a new test (a warning is shown if the hashes
of the implicitly excluded tests mismatch).