Fix #2523: A clean implementation of expandLongValue.
As the previous comment in `expandLongValue` said, it was a trick.
I thought that trick was at least *safe*, but it turns out it
wasn't in some corner cases, which were exposed by
`run/bitsets.scala`.

This commit fixes the issue with a clean implementation of
`expandLongValue` instead of a trick.