Consider module accessor a JSGetter
Redundant on 2.11, but required for the upcoming 2.12.0-RC1,
where the field refactoring synthesizes module accessors
after uncurry, so that their type history does not even
go back as far as uncurry, where a NullaryMethodType might
have been observed otherwise.

I tried splicing NMTs in the info history of modules,
but they already are pretty hacking (a module's term symbol
is converted into a method pretty late in the game, and it
would be risky to have them start with NMTs from namers).