Fix #2491: Consider all exported properties as externs for GCC.
The name clashes were caused by GCC renaming some Scala methods
or fields to a name that was also exported. Hence, a Scala class
could have, as instance member, both a (renamed) Scala
method/field and a (non-renamed) exported method or property.

To prevent this, we list all exported methods and properties in
the linking unit, and we feed them as externs to GCC. This forces
GCC to avoid those names when renaming Scala methods and fields.