Fix #2552: Store generalized "loading spec" for native things in IR.
Previously, only native JS classes stored in the IR the means to
load their constructor. This was stored as the `jsName` field, which
was the fully qualified name of the class constructor, starting from
the global object.

This information was notably used by the `LoadJSConstructor` IR node,
allowing that node to be valid both for native JS classes and
Scala.js-defined JS classes.

Native JS objects did not have a similar information, which caused
`LoadJSModule` to be valid only for Scala.js-defined JS objects.
Therefore, for native JS objects, the "desugaring" of `LoadJSModule`
that would logically happen at link-time was done by the compiler.
This meant that changing the `@JSName` of a native JS object was
not a binary compatible change, as recorded in issue #2552.

This commit fixes this issue, and in the process generalizes the
stored information as "loading spec". Loading specs apply both to
native JS classes and objects. They are a declarative specification
of how a class constructor or an object should be loaded by
`LoadJSConstructor` and `LoadJSModule`, respectively.

Instead of storing the unstructured fully qualified name, a loading
spec is an ADT with, currently, only one possible case, which is
`Global(path: List[String])`. In the future, other loading spec
kinds will be added, such as imports from ES6 modules (see #2175).

Since now both classes and objects have a loading spec, we cannot
use `jsName.isDefined` to identify native JS classes anymore (which
is necessary for IR checks and when emitting the run-time
`isInstance` reflective functions for raw JS types). To recover this
distinction in the IR, we split `ClassKind.RawJSType` into three:
`NativeJSClass`, `NativeJSModuleClass` and `AbstractJSType` (the
latter being used for JS traits). With this information, we can
check that `LoadJSConstructor` is valid for `NativeJSClass` and that
`LoadJSModule` is valid for `NativeJSModuleClass`.

We add a deserialization hack to turn `RawJSType` from the 0.6.8
format into `NativeJSClass` if there was a `jsName` (which we also
turn into a loading spec). There is a hack on top of the hack,
though. Since `jsName` is only visible in the tree, we cannot
apply this hack to the `kind` stored in the infos. Therefore, the
`kind` in the infos may be inconsistent with that in the tree.
This is only a slight issue for the `Analyzer`, which therefore has
another hack to tolerate an `AbstractJSType` where it expects a
`NativeJSClass`.

Note that native JS *objects* compiled with the 0.6.8 format will
be deserialized as `AbstractJSType`, which means that
`LoadJSModule` cannot be used on them. This means that even the new
compiler must eagerly desugar `LoadJSModule` when loading native JS
modules compiled by an older compiler, as it used to always do.
Since there is no way to otherwise tell whether an object was
compiled with a recent compiler or not, we add a marker annotation
`@HasJSNativeLoadSpec` to native JS objects compiled with the 0.6.13
format or later. For those, we can issue a `LoadJSModule` node.

In summary, this commit provides:

- A fix for #2552 for objects compiled with 0.6.13 or later.
- A generalized loading spec which can be extended to ES6 imports
  in the future.