Always use Node.js in partest, instead of Rhino in noOpt.
Historically, we used Rhino to execute the partest test suite
in --noOpt. This hasn't been necessary for a while, but we had
kept it that way to reduce the likelihood of breakage at the
time.

With the deprecation of the Rhino runner, it is time to forget
that dark past, and always use Node.js to run partest, even in
noOpt.