Tone down the number of combinations in CI.
Our resource consumption of CI has grown much too intense, to a
point where it is slowing us down, as PR testing can be delayed
by several hours.

This commit "strategically" reduces the number of combinations
that are tested on CI. Basically, we assume 2.10.2, 2.11.8 and
2.12.0-RC1 combined with JDK8 only as the "default", "base"
combination. From that combination, we vary all the other
parameters independently, but we do not combine several axes
together. One exception is for sbt-plugin related things, whose
"base" combination is 2.10.6 with all the JDKs.

To make sure that the `testSuite` (which contains JDK7-only and
JDK8-only parts) runs on JDK6 and 7 in `pr`, one `testSuite`
run (in the default config, i.e., Node.js + fastOpt) is added
to the `main` task.

Partest fastOpt moves from `nightly` to `pr`. The tests in `pr`
have become slow enough that two partest tasks (for 2.11.8 and
2.12.0-RC1) will not kill it. And this is the one thing from
`nightly` that often breaks, so it is worth having in `pr`.