Properly incrementalize the Emitter using knowledge queries.
When a paper's terminology makes its way into a production
codebase ...

The `ScalaJSClassEmitter` and `JSDesugaring` do not have access to
the `linkingUnit` anymore. All their requests for global knowledge
go through a `GlobalKnowledge` interface, which contains the
knowledge queries. The `Emitter` contains a subsystem
`KnowledgeGuardian`, which, as its name implies, is responsible for
guarding global knowledge. It tracks the dependencies between
arbitrary `Invalidatable`s and the knowledge queries they perform.
The `Invalidatable`s are `ClassCache` and `MethodCache`, which also
implement `GlobalKnowledge` by delegating queries to the
`KnowledgeGuardian`.

This fixes some corner case bugs in the incremental behavior of the
`Emitter`, the most significant one being changes in the `@JSName`
of a native JS class or object. Previously, users of
`LoadJSConstructor` (resp. `LoadJSModule`) would not be properly
re-desugared when the `@JSName` of a target class (resp. module)
was changed.

The new architecture can now easily be extended with new kinds of
global knowledge without fear of breaking incrementality.