Fix #2613: Replace `extends js.GlobalScope` by `@JSGlobalScope`.
This prompted a lot of refactoring in `PrepJSInterop` to keep
things DRY. In particular, all annotations related to the JS
native load specs (i.e., `@JSName`, `@JSImport` and
`@JSGlobalScope`) are now handled in (conceptually) the same way.
Some discrepancies still exist to account for combinations that
used not to be checked, as they need to warn instead of error out.