Fix #2625: Fix clash of outer pointers in 2.12.0-RC2 onwards.
The root cause of the bug #2625 which surfaces in 2.12.0 is
actually #2382. We cannot fix #2382 without breaking backwards
binary compatibility in general. However, we can fix it only
for 2.12.0-RC2 onwards, as RCs of Scala (and the upcoming final)
have separate binary ecosystems anyway.

So this commit fixes #2382 but only for 2.12.0-RC2 onwards. The
fix is trivial: name mangle all outer pointers in the same way
as if they were private, i.e., with the number of class ancestors
of its owner.