Fix #1381: Add top-level exports
The new annotation @JSExportTopLevel allows modules to export their
methods to the top-level of the export scope, so they can be called
directly. For example:

    object MyApp {
      @JSExportTopLevel("main")
      def main(): Unit = ???
    }

Assuming that the export namespace is the `window` object, `MyApp.main`
can now be invoked as follows in JavaScript:

    window.main()

As opposed to (if one assumes `MyApp` and `main` to be exported):

    window.MyApp().main()

Overloading is supported from within the same module. Exporting to the
same name (irrespective of the export type) from different modules will
cause a link time error.

Current restrictions of top-level exports:

- Properties (or fields) cannot be exported.
- The fully qualified name of the export needs to be given.