Add an arithmetic simplification in optimizer for >>>.
    (c1 & x) >>> c2 --> (c1 >>> c2) & (x >>> c2)

And the same with `|` and `^`. Note that `c1 >>> c2` is still
constant, so we are not actually creating more operations.
Moreover, the value of `c1 >>> c2` is textually shorter-or-equal
than `c1`, so at worst we keep the same character length, and
sometimes slightly improve it.

The benefit, though, happens if `c1 >>> c2` evaluates to the
absorbing element of the bitwise operation, everything disappears.
This happens notably for expressions like

    (0xffff & x) >>> 16

which will be produced by the improved `Long` multiplication of the
following commit, in some cases.