Shorter and faster `Long` multiplication.
The key insight was that we do not need to compute the `hi`
products by 16-bit components. The `hi` part is basically

    a.lo*b.hi + a.hi*b.hi + carry_from_lo_*

so only the carry from multiplying `a.lo*b.lo` needs to be
computed by 16-bit components. The resulting algorithm is
significantly shorter, and we can afford to completely inline
it.

Subsequently, we can use the 16-bit products computed for the
hi part to also compute the lo part, removing one multiplication.

The fact that we can now inline the whole algorithm allows it to
benefit from call-site optimizations on the underlying `Int`
arithmetics. This mostly happens when one side is constant and/or
the `hi` part of operands is 0. The whole thing collapses
particularly well when `a` is a constant in the range
`0 <= a <= 0xffff`, which happens for mundane things like `10 * x`.

This improved implementation is significantly faster. A benchmark
based on `BigInteger.multiply` exhibits > 2x improvement.