25% faster Long.toString() for large Longs (>= 2^53).
For large `Long` values, `toString()` divides once by 10^9, then
converts the quotient and remainder to strings and concatenates
the results.

Previously, the quotient and remainder were returned in a tuple
from `unsignedDivModHelper()`. Now, we let the latter perform
the final step of conversion to string, which avoids the tuple
allocation. According to micro-benchmarks, this optimization
shaved off 25% of the run-time of `Long.toString()` for large
`Long` values.