Add an annotation `@JSOptional` for optional members in JS types.
In a JS type, and particularly in a Scala.js-defined JS type, vals,
vars and defs without parentheses can be annotated with
`@JSOptional`. In that case, they must be concrete, and their rhs
must be `js.undefined`. This is also valid in a Scala.js-defined JS
trait, and is the only way to define a concrete member in such a
trait.

An `@JSOptional` member does not create a field nor getter in the
object or class. Instead, it gets a value of `undefined` by virtue
of JavaScript using `undefined` for missing fields. In case of a
`var`, the field can be created later by explicit assignment. For
`val`s and `def`s, it can be created in subclasses with an
`override val` or `override def`.

This is particularly useful to model "configuration objects" as
used in many JavaScript APIs. A Scala.js-defined JS trait with
`JSOptional` fields can define a typed API for the configuration
object. It is then easy to create a configuration object with only
a few fields set by defining `override val`s. Other fields stay
optional (by inheritance) and are not created on the resulting
JavaScript object.