Allow concrete val/var/defs in JS traits with `= js.undefined`.
Normally, term members in Scala.js-defined JS traits must be
abstract. This commit relaxes the restriction to allow vals, vars
and defs without parentheses to be concrete, if their right-hand-
side is `js.undefined`.

Such fields and methods are not actually created on
Scala.js-defined JS classes and objects that extend those traits.
Instead, they get a value of `undefined` by virtue of JavaScript
using `undefined` for missing fields. In case of a `var`, the
field can be created later by explicit assignment. For `val`s and
`def`s, it can be created in subclasses with an `override val` or
`override def`.

This is particularly useful to model "configuration objects" as
used in many JavaScript APIs. A Scala.js-defined JS trait with
concrete `js.undefined` fields can define a typed API for the
configuration object. It is then easy to create a configuration
object with only a few fields set by defining `override val`s.
Other fields and are not created on the resulting JavaScript
object.