Disallow other bad @JSBracketAccess params
Note that this is technically speaking a backwards compatibility
breaking change. It was possible to define repeated or default arguments
in @JSBracketAccess methods, as long as the invocation site had one or
two parameters. However, all of the (working) use cases this breaks can
be replaced by overloads in a fully source and binary compatible
manner.