Fix #2631: Gracefully fail if UndefinedParam slips
In Scala, default parameters are resolved at call site. Therefore, it is
valid for a macro or a compiler plugin to move them around or even
completely decouple from the function call they originated from.

However, for Scala.js' facade types, this is not the case, mainly
because in JavaScript, default parameters are resolved at definition
site. Macros and compiler plugins must (to some extent) be aware of this
restriction when running together with the Scala.js compiler plugin.

With this commit, we gracefully report if we cannot compile a tree we
received due to this restriction. Before this fix, the compiler emitted
invalid IR.