Add a helper `js.defined(x)` to drive type inference in some cases.
When the expected type of an expression `x` is of the form
`js.UndefOr[A]`, but an implicit conversion or a SAM conversion is
required to type `x` as an `A`, the normal implicit conversion to
`js.UndefOr[A]` does not work.

We can use the helper `js.defined(x)` to drive Scala's type
inference to apply the implicit conversion or the SAM conversion.
This is particularly relevant for SAM conversion, for which even

  def any2undefOrA[A, B](a: A)(implicit ev: A => B): UndefOr[B]

would not be enough to allow parameter type inference anyway.

This is particularly important for configuration objects with
function fields.