Allow to export module methods as static in their companion class.
This implements the first part of the proposal #2702, i.e.,
exporting methods as static.

This commit introduces a new annotation `@JSExportStatic`. It can
be used on (non-accessor) methods in Scala objects whose companion
class is a Scala.js-defined JS class. The annotation causes the
method to exported as a static method of the companion class.

The mechanism is conceptually (and technically) quite close to
`@JSExportTopLevel`, where the export is nested in a static class
namespace instead of being at the top-level.