Fix #1902: Export module fields as static in their companion class.
This commit implements the second part of the proposal #2702, i.e.,
exporting fields as static.

The `@JSExportStatic` annotation can now be used on `val`s and
`var`s in addition to methods. This causes the specified field to
be exported (and in fact stored) as a static field in the companion
class of the enclosing module. Since this affects the storage
location of the field, it cannot be used twice (with different
export names) on the same field, as that would require the field
to be stored in two different locations.