Export fields at the top-level with `@JSExportTopLevel`.
This commit implements the third part of the proposal #2702, i.e.,
exporting fields at the top-level.

The `@JSExportTopLevel` annotation can now be used on `val`s and
`var`s in addition to methods. This causes the specified field to
be stored as a static field in the IR (which becomes a top-level
var in JS) and exported at the top-level. Field exports are
read-only views, as is the case in ECMAScript 2015 modules, so
JS code cannot modify exported vars, but can observe changes to
it done by Scala code.