Fix #2715: Emit object literals for new anonymous JS class.
When instantiating an anonymous JS class, we actually instantiate
the parent class. If that parent class happens to be `js.Object`,
and there is no formal argument to the constructor, we can use an
empty object literal `{}` instead.

This makes the compiler behavior consistent with a direct
`new js.Object()`, for which we already applied this optimization.
This allows Closure to fuse it with subsequent field assignments,
eventually producing a direct object literal with fields.