Fix #1997: Add support for js.Symbols in @JSName annotations.
The `@JSName` annotation could previously be used to specify the
JavaScript *string* name of a property/method, independently of
its Scala name. This can be used both in native JS classes as well
as Scala.js-defined JS classes.

This commit extends the capability of `@JSName` to admit a (static,
stable) reference to a `js.Symbol`. This applies to both native
and Scala.js-defined JS types.

At call site, the semantics and compilation scheme are
straightforward: it is a `JSBracketSelect` whose `item` part is
a call to the symbol accessor. Since it is static by construction,
we can call it from anywhere.

At definition site for native JS types, there is nothing to do
(besides validity checks).

At definition site for Scala.js-defined JS classes, things get
tricky. We can now have "exported" class members whose name is
not a constant string anymore. This requires to extend the IR with
`ComputedName`s (a concept that exists in ECMAScript 2015). A
computed name is an arbitrary expression `Tree` that is evaluated
to obtain the dynamic name of a property. This complicates the
Emitter somewhat, but not that much because Scala.js-defined JS
classes are created lazily, so we can execute arbitrary
user-defined code at class-definition time.

For overloading resolution, `ComputedName`s have a `logicalName`,
derived from the fully qualified name of the symbol accessors.
All `ComputedName`'d properties/methods with the same logical
name (i.e., referencing the same static stable `js.Symbol`) are
considered to be overloaded. This has an unavoidable caveat: if
two `ComputedName`s in the same class have a different logical
name, but actually evaluate to the same symbol, things will go
wrong at run-time.

`JSObjectConstr` is enhanced to support `ComputedName`s as well,
for consistency. It is also consistent with Object Initializers
in ECMAScript 2015. This actually complicates their treatment,
mostly because of evaluation order concerns.

Exports (in Scala classes and at the top-level) cannot use
symbolic names. This is not expected to change in the future.

There is one implementation restriction: `@JSName` with a symbol
is not supported on native JS classes and objects that are nested
within a native JS object. There is no language reason to
disallow this, but the implementation is surprisingly difficult,
due to the way scalac flattens such entities.