Fix missing __exportedInits info for top-level field exports.
In practice, it wasn't an issue, because a) a top-level field
export would always be accompanied by a static initializer,
ensuring the class would be kept, and b) fields are never dce'ed
away at the moment.

It would have become an issue in the future, though, if/when the
analyzer gets smarter about fields.