Prepare the testing interface to transition to reflective instantiation.
This commit makes all the necessary changes to the testing interface
so that it will be able to cope with reflective instantiation-based
test loading. This means using `@EnableReflectiveInstantiation`
instead of `@JSExportDescendantClasses`/`Objects` on both
`sbt.testing.Framework` and test classes.

This commit does *not*, however, already change the annotations.
This ensures that the changes are "binary-compatible", in the sense
that a new sbt plugin and (possibly) a new testing interface are
able to run testing frameworks designed with the old export-based
system.