Fix #2734: Allow @JSExportTopLevel on classes and objects.
This commit implements the Proposal #2734. The annotation
`@JSExportTopLevel` can now be used on static classes and objects
(both Scala classes/objects and Scala.js-defined JS
classes/objects) to export them at the top-level.

For objects, the difference wrt. `@JSExport` is that the module
instance itself is exported, rather than the module accessor.

For classes, this is equivalent to exporting them via `@JSExport`.
We include the possibility to be consistent with objects, and with
the prospect of deprecating and removing `@JSExport` on classes
and objects altogether.