Fix #2265 and #2741: Look for repeated args in all parameter lists.
Usually, in the compiler back-end, we use `methodTpe.params` to
read the list of all parameters. This works because multiple
parameter lists are flattened. The code that optimizes varargs as
`new js.WrappedArray(js.Array(...))` was reading parameters right
after `typer` in the type history. It has to use `.paramss.flatten`
in that case, because `.params` represents only the first parameter
list.