Fix #2532: Emit "private" field access for Java-defined fields.
Java-defined fields are now always accessed as if they were private.
This is necessary because they are defined as private by our .scala
source files, but they are considered `!isPrivate` at use site,
since their symbols come from Java-emitted .class files.
Fortunately, we can easily detect those as `isJavaDefined`.

This includes fields of Ref types (IntRef, ObjectRef, etc.) which
were previously special-cased at use-site. This commit generalizes
the idea to all fields that are Java-defined (as per the .class
files).

Caveat: because of this, changing the length of the superclass
chain of a Java-defined class is a binary incompatible change.

This also fixes #2737, which was directly caused by #2532.