Fix #2752: Survive in `test` if there is no testing framework.
The `test` task should trivially succeed if there is no testing
framework on the classpath. However,
f46050f245bed6c318e2f2855a811ea6afb5d0cd broke this scenario, by
requiring an export from the testing interface to be present.

This commit fixes the scenario to tolerate the absence of that
export. In that case, no candidate testing framework is reported
to be available.

We add `library/test` to the CI matrix as a test for this scenario,
as the `library` project does not and will not contain any tests.