Fix #2762: Remove obsolete Node.js flags for rest and spread.
Node.js 6.x, which is now installed in the CI servers, does not
support the `--harmony-rest-parameters` and `--harmony-spreadcalls`
flags anymore. Instead, it supports those features by default.

This commit removes those flags when running with Node.js. This
fixes the issue if a recent Node.js is used.

Caveat: this fix will *break* runs in ES6 mode with an older
Node.js. I do not see an easy way to correctly run on both old and
new versions. Since this is an internal issue anyway, I don't think
it's worth complicating the build with logic to detect the version
of Node.js or whatnot.