Fix #2755: Fix array out of bounds in BigDecimal.divide(scale).
There were computations based on `LongTenPows` that would access
indices before testing whether they are valid or not. The fix is
to delay those accesses after we've made sure that indices are
valid.

In the process, I have rearranged the computation of `diffScale`
to convert it to an `Int` earlier, avoid some useless `Long`
operations.