Make the union type understand variance for 1-arg type constructors.
The subtyping rules of the union type, materialized by implicit
`Evidence`s, are enhanced with support for variance in type
constructors of 1 argument.

Concretely, this means that `List[Int]` can now be assigned to a
`List[Int | String]`, and that a `Consumer[Int | String]` can be
assigned to a `Consumer[Int]` (for some `Consumer[-A]`).