Add user-defined module initializers to the linker.
A module initializer is a piece of code that is executed when
the module produced for a linked Scala.js application is loaded.

This commit adds an sbt task of type `Seq[ModuleInitializer]`,
which allows the user to define custom entry points to their
module. Example:

  scalaJSModuleInitializers += {
    org.scalajs.core.tools.linker.ModuleInitializer.mainMethod(
        "fully.qualified.ObjectName", "main")
  }

Module initializers can be specified `in Compile`, in which case
they are not executed during tests.

This fills a gap in our interoperability criterium, as previously
there was no way that user-defined code could be executed at
module load time (without hacking around it by using an exported
static field that is not actually used).

This is also intended to provide a basis for a better replacement
to launcher files, but such work is not done in this commit.