Add scalaJSUseMainModuleInitializer, to use an initializer for the mainClass.
When the setting `scalaJSUseMainModuleInitializer` is set to true,
a module initializer will be derived from `mainClass` (which is
itself derived from discovered `js.JSApp`s in the program). This
causes a call to the main method of a main class to be built in
the fastOptJS/fullOptJS output.

Consequently, `run` is adapted to know about
`scalaJSUseMainModuleInitializer`, so that it doesn't execute the
main method twice.

The setting is mutually exclusive with `persistLauncher`, as it
does not make sense to write a launcher file when the main .js
file already executes the main method.

In fact, `scalaJSUseMainModuleInitializer` is intended to supersede
launcher files and `persistLauncher`.

This also fixes #2199.