Deprecate top-level @JSExport, and @JSExportDescendent{Classes,Objects}.
Top-level `@JSExport`s (on classes and objects) should be replaced
by `@JSExportTopLevel`. For objects, this is a change of semantics,
as the object itself is exported, rather than a 0-arg function
returning the object. It is still possible to achieve the same
behavior as before by explicitly exporting a 0-arg method returning
the object with `@JSExportTopLevel`, so we do not lose any
expressiveness.

`@JSExportDescendentClasses` and `@JSExportDescendentObjects`
should not have any use cases left, now that reflective
instantiation and module loading is properly supported through
`@EnableReflectiveInstantiation` and `scala.scalajs.reflect.Reflect`.