Do not include java.math tests when using PhantomJS.
This reduces the pressure on PhantomJS in terms of mass of code
in the test suite.

It seems we reached the limits of what PhantomJS can handle in
terms of code mass. Since PhantomJS support is due to be moved to
a separate repository in 1.0.0, the easiest way to fix this is to
reduce the pressure on PhantomJS. We therefore remove the tests of
java.math (BigInteger and BigDecimal) when running with PhantomJS.
These tests are well isolated, and the less likely to have
environmental differences.