Demand `@JSGlobal` for non-`@JSImport` JS native classes and objects.
Previously, a top-level `@js.native` class/object was loaded from
the global scope by default, either with an explicit name through
`@JSName` or with an implicit name derived from its Scala name.

With this commit, such a native class/object should be declared
with an `@JSGlobal` annotation instead (either with or without an
explicit name).

For example,

  @js.native class Foo extends js.Object
  @js.native @JSName("Foobaz") object Bar extends js.Object

would now be written as

  @js.native @JSGlobal class Foo extends js.Object
  @js.native @JSGlobal("Foobaz") object Bar extends js.Object

There are two main reasons for such a change.

First, it brings loading from the global scope syntactically on
par with importing from a module. As the JS eco-system moves
towards modules, the syntactic bias towards globally scoped things
is not relevant anymore.

Second, it removes the overloaded meaning of `@JSName`. Indeed,
this annotation has significantly different meanings whether it
is attached to a top-level native class/object versus a member of
a JS type. For example, a `.` has a special meaning for top-level
native class/object, but not for a member of a JS type. The
changes in this commit close #2615 by leaving only the meaning
where `.` is not special, and therefore need not be forbidden.

Since this is a fairly disruptive change, the deprecation warnings
can be suppressed with the option
`-P:scalajs:suppressMissingJSGlobalDeprecations` during the 0.6.x
cycle, to ease migration.