Deprecate private members of native JS classes/traits/objects.
Historically, our compiler has tolerated `private`, `private[this]`
and `private[Scope]` members in native JS classes, traits, and
objects. Once compiled, it would consider that they are in fact
public, with non-mangled names.

This is at odds with how private members of Scala.js-defined JS
types are handled at use site, which means there are different
semantics at use site between native versus Scala.js-defined types,
which is bad.

This commit deprecates definitions of private members in native
JS types. We still accept `private[this]` constructors, because
they are sometimes necessary to have a primary constructor that
other constructors delegate to. `private[this]` constructors are
fine, because they cannot be called from anywhere but the
secondary constructors. Note that even a `private` constructor
can be called from the companion object, which can be non-native,
so this is not allowed.