Fix #2799: Don't use `equals` for comparing java.lang.Double/Float.
This is essentially a port of the fix for scala-dev#329, which was
implemented in https://github.com/scala/scala/pull/5761.

To remain bug-compatible with the JVM, we explicitly look at the
version of the Scala compiler that we run on to determine whether
or not the fix should be applied.