Do not automatically select JSDOMNodeJSEnv if requiresDOM is true.
This removes the interaction between `jsDependencies` on the one
hand, and `JSDOMNodeJSEnv` on the other hand. This will be required
so that we can move those two segments of the sbt plugin in
separate repositories and sbt plugins (#2841, #2842).

Note that this makes `requiresDOM`, and therefore
`jsDependencies += RuntimeDOM`, completely useless; they do not
have any effect anymore. They should probably be deprecated/stubbed
out in the `jsDependencies` plugin.