Fix #2780: Reach all methods of j.l.Integer before the optimizer.
This was already partially fixed in
10997e9b04eb1a82225f85975073a64c09982448, which was a fix for
`compareTo(Short)` were affected. #2780 shows that all methods of
`java.lang.Integer` are affected, in fact, so this commit extends
the existing fix to all methods.

Note that the methods still disappear after the refiner if they
are not needed, and anyway they are all hijacked as helpers, so
the quality of DCE is not affected by this fix.

There is no additional test, because the warnings could already
be observed before, but as warnings, they did not fail the build.