Actually use `enablePlugins(ScalaJSPlugin)` inside our own build.
We previously avoided that not to have so-called "ecosystem"
settings, e.g., library dependencies on our own artifacts. It is
however equally easy to cancel out those settings we do not want
in a custom `AutoPlugin` specific to the build.

This allows us to use a genuine `enablePlugins`, which in turn
will allow us to add more `enablePlugins` on plugins that depend
on `ScalaJSPlugin` without breaking everything. For example, we
do this with `PhantomJSEnvPlugin`.